refactor: separate update in smaller files
- current month
- selected date
